requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
streamlit
python-dotenv
schedule
pytest
pandas